refactor: extract time into own component and reorganize directory structure
